requirements.txt
# pip
absl-py==0.15.0
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
chardet==5.0.0
charset-normalizer==2.0.12
clang==5.0
colorama==0.4.5
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
easydict==1.10
entrypoints==0.3
ffmpeg==1.4
ffmpy==0.3.0
filetype==1.2.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
h5py==3.1.0
idna==3.4
imageio==2.15.0
imageio-ffmpeg==0.4.8
importlib-metadata==4.8.3
importlib-resources==5.4.0
imutil==0.3.4
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.17.0
Jinja2==3.0.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
moviepy==1.0.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.3
numpy==1.19.2
oauthlib==3.2.2
olefile==0.46
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==6.2.2
pip==21.2.2
proglog==0.1.10
prometheus-client==0.13.1
prompt-toolkit==3.0.20
protobuf==3.19.6
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.9
pypiwin32==223
pyrsistent==0.17.3
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.1.1
pywin32==305
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==25.0.1
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==58.0.4
six==1.16.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.5.0
threadpoolctl==2.2.0
tifffile==2020.9.3
torch==1.8.2
torchaudio==0.8.2
torchvision==0.9.2
tornado==6.1
tqdm==4.64.1
tracker==0.1.1
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.15
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
wheel==0.37.1
widgetsnbextension==3.5.1
win10toast==0.9
wincertstore==0.2
wrapt==1.12.1
zipp==3.6.0

#anaconda3
absl-py==0.15.0
argon2-cffi==20.1.0
astunparse==1.6.3
async_generator==1.10
attrs==21.4.0
backcall==0.2.0
blas==1.0
bleach==4.1.0
ca-certificates==2023.01.10
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
chardet==5.0.0
charset-normalizer==2.0.12
clang==5.0
colorama==0.4.5
cudatoolkit==11.1.1
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
easydict==1.10
entrypoints==0.3
ffmpeg==1.4
ffmpy==0.3.0
filetype==1.2.0
flatbuffers==1.12
freetype==2.12.1
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
h5py==3.1.0
icc_rt==2022.1.0
icu==58.2
idna==3.4
imageio==2.15.0
imageio-ffmpeg==0.4.8
importlib-metadata==4.8.3
importlib-resources==5.4.0
importlib_metadata==4.8.1
imutil==0.3.4
intel-openmp==2023.0.0
ipykernel==5.3.4
ipython==7.16.1
ipython_genutils==0.2.0
ipywidgets==7.6.5
jedi==0.17.0
jinja2==3.0.3
joblib==1.0.1
jpeg==9b
jsonschema==3.2.0
jupyter==1.0.0
jupyter_client==7.1.2
jupyter_console==6.4.3
jupyter_core==4.8.1
jupyterlab_pygments==0.1.2
jupyterlab_widgets==1.0.0
keras==2.6.0
keras-preprocessing==1.1.2
kiwisolver==1.3.1
lerc==3.0
libdeflate==1.17
libpng==1.6.39
libtiff==4.5.0
libuv==1.44.2
lz4-c==1.9.4
m2w64-gcc-libgfortran==5.3.0
m2w64-gcc-libs==5.3.0
m2w64-gcc-libs-core==5.3.0
m2w64-gmp==6.1.0
m2w64-libwinpthread-git==5.0.0.4634.697f757
markdown==3.3.7
markupsafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mkl==2020.2
mkl-service==2.3.0
mkl_fft==1.3.0
mkl_random==1.1.1
moviepy==1.0.3
msys2-conda-epoch==20160418
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
ninja==1.10.2
ninja-base==1.10.2
notebook==6.4.3
numpy==1.19.2
numpy-base==1.19.2
oauthlib==3.2.2
olefile==0.46
opencv-python==4.6.0.66
openssl==1.1.1t
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
pandoc==2.12
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pillow==6.2.2
pip==21.2.2
proglog==0.1.10
prometheus_client==0.13.1
prompt-toolkit==3.0.20
prompt_toolkit==3.0.20
protobuf==3.19.6
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygments==2.11.2
pyparsing==3.0.9
pypiwin32==223
pyqt==5.9.2
pyrsistent==0.17.3
python==3.6.13
python-dateutil==2.8.2
pytorch==1.8.2
pytz==2022.7.1
pywavelets==1.1.1
pywin32==305
pywinpty==0.5.7
pyyaml==5.4.1
pyzmq==25.0.1
qt==5.9.7
qtconsole==5.2.2
qtpy==2.0.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
send2trash==1.8.0
setuptools==58.0.4
sip==4.19.8
six==1.15.0
sqlite==3.41.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.5.0
threadpoolctl==2.2.0
tifffile==2020.9.3
tk==8.6.12
torchaudio==0.8.2
torchvision==0.9.2
tornado==6.1
tqdm==4.64.1
tracker==0.1.1
traitlets==4.3.3
typing-extensions==3.7.4.3
typing_extensions==4.1.1
urllib3==1.26.15
vc==14.2
vs2015_runtime==14.27.29016
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==2.0.3
wheel==0.37.1
widgetsnbextension==3.5.1
win10toast==0.9
wincertstore==0.2
winpty==0.4.3
wrapt==1.12.1
xz==5.2.10
zipp==3.6.0

# base
certifi
chardet
cycler
dataclasses
easydict
idna
imutil
kiwisolver
matplotlib
numpy
opencv-python
Pillow
pkg-resources
pyparsing
python-dateutil
pyyaml
requests
scipy
six
tqdm
typing-extensions
urllib3
pandas
seaborn